# Conflicts:
#	.gitignore
#	README.md
#	fpdb-aws/include/fpdb/aws/AWSConfig.h
#	fpdb-aws/src/AWSClient.cpp
#	fpdb-aws/src/ProfileAWSCredentialsProviderChain.cpp
#	fpdb-main/test/TestUtil.cpp

+++ fpdb-tuple/include/fpdb/tuple/Globals.h
@@ -15,7 +15,7 @@ namespace fpdb::tuple {
 /**
  * Default number of records per arrow chunk when processing record batches
  */
-inline size_t DefaultChunkSize = 100000;
+inline int64_t DefaultChunkSize = 100000;
 
 }
